[android] possible NPE on transfer removal

+++ android/src/com/frostwire/android/gui/transfers/UIBittorrentDownload.java
@@ -1,19 +1,18 @@
 /*
  * Created by Angel Leon (@gubatron), Alden Torres (aldenml)
- * Copyright (c) 2011-2017, FrostWire(R). All rights reserved.
+ * Copyright (c) 2011-2020, FrostWire(R). All rights reserved.
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package com.frostwire.android.gui.transfers;
@@ -345,13 +344,15 @@ public final class UIBittorrentDownload implements BittorrentDownload {
     }
 
     private void finalCleanup(Set<File> incompleteFiles) {
-        for (File f : incompleteFiles) {
-            try {
-                if (f.exists() && !f.delete()) {
+        if (incompleteFiles != null) {
+            for (File f : incompleteFiles) {
+                try {
+                    if (f.exists() && !f.delete()) {
+                        LOG.info("Can't delete file: " + f);
+                    }
+                } catch (Throwable e) {
                     LOG.info("Can't delete file: " + f);
                 }
-            } catch (Throwable e) {
-                LOG.info("Can't delete file: " + f);
             }
         }
 
@@ -429,7 +430,9 @@ public final class UIBittorrentDownload implements BittorrentDownload {
         return l;
     }
 
-    /** Downloads the first 10mb or 30% of the file sequentially */
+    /**
+     * Downloads the first 10mb or 30% of the file sequentially
+     */
     private void checkSequentialDownload() {
         BTDownloadItem item = getFirstBiggestItem();
